M7 infrastructure split: per-app and platform stacks to canonical homes (closes #250)
- Move 8 platform stacks infrastructure/lib/platform/ → platform/infrastructure/
- Move 2 SA stacks infrastructure/lib/stock-analyser/ → apps/stock-analyser/infrastructure/
- Move 3 BT stacks infrastructure/lib/budget-tracker/ → apps/budget-tracker/infrastructure/
- Update infrastructure/bin/app.ts import paths
- Fix Lambda path.join(__dirname, ...) in 6 stack files for new __dirname values
- Update infrastructure/tsconfig.json includes + add paths override for aws-cdk-lib dedup
- Update deploy workflow path filters (deploy-platform, deploy-stock-analyser, deploy-budget-tracker)
- Delete apps/web/ and apps/web-vite-backup/; remove !apps/web-vite-backup exclusion from pnpm-workspace.yaml
- Update normative docs: MONOREPO.md, docs/architecture/cdk.md, docs/architecture/inventory.md, CONTRIBUTING.md §3.7, PLAN.md, both app CLAUDE.md files
