+++ apps/budget-tracker/infrastructure/budget-tracker-tables-stack.ts
