+++ platform/infrastructure/cognito-hosted-ui-css.ts
