+++ platform/infrastructure/network-stack.ts
