+++ platform/infrastructure/platform-tables-stack.ts
