+++ platform/infrastructure/storage-stack.ts
